wip testing with browserstack
The goal is to use browserstack to test cookie and XSRF policies
in various browsers. At this time the test is not expected to pass,
but it is expected to print
proxy received RequestStartLine(method='GET', path='http://example532532.com:40533/login', version='HTTP/1.1')

Instead, it's failing without ever reaching the proxy.

Repro steps:
1. Prerequisites: Python 3.10, tested on linux. 3.11 will probably
   work too. `pip install selenium browserstack-local`
2. From repo root, run `PYTHONPATH=. python3 maint/test/browser/browser_test.py`